Fix Core Data Model and related NSManagedObjects.
This is a VERY important commit. Understand what I did here. In order for an app to run smoothly, the models must be set up correctly, or else you're just going to fight with it.

Notice how the List entity has one attribute called "listName" (the name of the list), and a relationship called "items". "Items" is plural because one LIST has many ITEMS. If you have a List object, it has Items associated with it. I also fixed the relationship, which was previously one-to-one list to item, to one-to-many list to items.

The Item entity has an "itemName" attribute (the name of the item), and a relationship called "list". "List" is singular because an ITEM belongs to a LIST. It can't belong to two lists, just one list.