After my last commit I realized that there was a major disconnect in what the two of us were naming things.
I had assumed that that the first view controller is to hold Lists, not Items. In my head, many Items make up a List. So it would follow that on the first view controller you create a List (such as "Groceries"; or "Christmas Presents"), and in the second view controller you add Items to each list (such as "Milk", "Eggs", and "Bread"; or "Playstation", "Books", and "Clothes").

With that in mind, I went through the app and converted / replaced everything so that the view controllers manipulate the data models as I assumed they would.

My apologies for this misunderstanding. Things should be clearer from here on out.

If you check out this commit you'll still see that all "Items" still appear for whichever "List" is selected on the first view controller. That is intended, I'll address that issue in the next commit.